add client component for blog page
